backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
shap
google-genai
pydantic
python-dotenv
